# Minor Blitting optimizations - Bitmaps 3 of N