--- a/src/bitmap.cpp
+++ b/src/bitmap.cpp
@@ -935,21 +935,29 @@ void Bitmap::BlendBlit(int x, int y, Bitmap const& src, Rect const& src_rect, co
 		return;
 	}
 
-	if (&src != this)
+	pixman_image_t* mask = nullptr;
+	int mask_x = 0;
+	int mask_y = 0;
+
+	if (&src != this) {
 		pixman_image_composite32(src.GetOperator(),
 								 src.bitmap.get(), nullptr, bitmap.get(),
 								 src_rect.x, src_rect.y,
 								 0, 0,
 								 x, y,
 								 src_rect.width, src_rect.height);
+		mask = src.bitmap.get();
+		mask_x = src_rect.x;
+		mask_y = src_rect.y;
+	}
 
 	pixman_color_t tcolor = PixmanColor(color);
 	auto timage = PixmanImagePtr{ pixman_image_create_solid_fill(&tcolor) };
 
 	pixman_image_composite32(PIXMAN_OP_OVER,
-							 timage.get(), src.bitmap.get(), bitmap.get(),
+							 timage.get(), mask, bitmap.get(),
 							 0, 0,
-							 src_rect.x, src_rect.y,
+							 mask_x, mask_y,
 							 x, y,
 							 src_rect.width, src_rect.height);
 }