Depends on #1989