@fmatthew5876 2019-12-17
Performance differences in flip blitting:

Master
```
---------------------------------------------------
Benchmark            Time           CPU Iterations
---------------------------------------------------
BM_FlipBlit      71557 ns      71555 ns       9793
BM_Flip          93533 ns      93530 ns       7500
```

This PR
```
---------------------------------------------------
Benchmark            Time           CPU Iterations
---------------------------------------------------
BM_FlipBlit      71263 ns      69876 ns      10061
BM_Flip          59174 ns      57976 ns      12220
```