@fmatthew5876 2020-01-10
These are small improvements to some of our blitting routines. Some of them I looked into pixman code and found that the functions we were calling were just wrappers around other functions and simplified them. 

The only one that improves by a lot is `Flip()`. The rest barely register on the benchmark because all the time is dominated by the actual composite operation.